Fix estimates for table 9.8 top panel
TWFE estimates from earlier in the chapter are not the same as those
from the DD comparison (which is why you get 0.0769 that changes to
0.069 for the bacon decomp estimate). I had thought this was a mistake
in the text (omitting the "7") but I was wrong.

Also added a verification that this twfe estimate is the sum of the
weighted 2x2 estimates.